+++ ktlint-plugin/src/main/kotlin/com/nbadal/ktlint/KtlintLogger.kt
@@ -1,11 +1,9 @@
 package com.nbadal.ktlint
 
-import com.intellij.openapi.diagnostic.DefaultLogger
+import org.jetbrains.kotlin.utils.PrintingLogger
 
 @Suppress("unused")
-class KtlintLogger(
-    qualifiedName: String?,
-) : DefaultLogger(qualifiedName ?: "ktlint-intellij-plugin") {
+class KtlintLogger : PrintingLogger(System.out) {
     fun debug(
         throwable: Throwable? = null,
         message: () -> String?,