+++ ktlint-plugin/src/main/kotlin/com/nbadal/ktlint/actions/FormatAction.kt
@@ -12,26 +12,34 @@ import com.intellij.psi.PsiManager
 import com.nbadal.ktlint.KtlintFeature.SHOW_MENU_OPTION_FORMAT_WITH_KTLINT
 import com.nbadal.ktlint.KtlintFileAutocorrectHandler
 import com.nbadal.ktlint.KtlintMode.DISTRACT_FREE
+import com.nbadal.ktlint.KtlintNotifier.KtlintNotificationGroup.DEFAULT
 import com.nbadal.ktlint.KtlintNotifier.notifyInformation
 import com.nbadal.ktlint.KtlintNotifier.notifyWarning
 import com.nbadal.ktlint.KtlintRuleEngineWrapper
 import com.nbadal.ktlint.actions.FormatAction.KtlintFormatContentIterator.BatchStatus.FILE_RELATED_ERROR
-import com.nbadal.ktlint.actions.FormatAction.KtlintFormatContentIterator.BatchStatus.PLUGIN_CONFIGURATION_ERROR
 import com.nbadal.ktlint.actions.FormatAction.KtlintFormatContentIterator.BatchStatus.SUCCESS
 import com.nbadal.ktlint.isEnabled
+import com.nbadal.ktlint.isKotlinFile
 import com.nbadal.ktlint.ktlintMode
 
 class FormatAction : AnAction() {
     override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT
 
     override fun update(event: AnActionEvent) {
-        val project = event.getData(CommonDataKeys.PROJECT) ?: return
-        val files = event.getData(CommonDataKeys.VIRTUAL_FILE_ARRAY) ?: return
-
-        event.presentation.apply {
-            isEnabledAndVisible = project.isEnabled(SHOW_MENU_OPTION_FORMAT_WITH_KTLINT) && files.isNotEmpty()
+        event.apply {
+            presentation.isEnabledAndVisible = showMenuOptionFormatWithKtlint() && isDirectoryOrKotlinFileSelected()
         }
     }
+
+    private fun AnActionEvent.showMenuOptionFormatWithKtlint() =
+        getData(CommonDataKeys.PROJECT)
+            ?.run { isEnabled(SHOW_MENU_OPTION_FORMAT_WITH_KTLINT) }
+            ?: false
+
+    private fun AnActionEvent.isDirectoryOrKotlinFileSelected() =
+        getData(CommonDataKeys.VIRTUAL_FILE_ARRAY)
+            ?.any { it.isDirectory || it.isKotlinFile() }
+            ?: false
 
     override fun actionPerformed(event: AnActionEvent) {
         val files = event.getData(CommonDataKeys.VIRTUAL_FILE_ARRAY) ?: return
@@ -57,6 +65,7 @@ class FormatAction : AnAction() {
         when (ktlintFormatContentIterator.status) {
             SUCCESS -> {
                 notifyInformation(
+                    notificationGroup = DEFAULT,
                     project = project,
                     title = "Format with Ktlint",
                     message = message,
@@ -65,14 +74,11 @@ class FormatAction : AnAction() {
 
             FILE_RELATED_ERROR -> {
                 notifyWarning(
+                    notificationGroup = DEFAULT,
                     project = project,
                     title = "Format with Ktlint",
                     message = message,
                 )
-            }
-
-            PLUGIN_CONFIGURATION_ERROR -> {
-                // Notification is already sent by plugin
             }
         }
     }
@@ -115,12 +121,6 @@ class FormatAction : AnAction() {
                     return true
                 }
 
-                KtlintRuleEngineWrapper.KtlintResult.Status.PLUGIN_CONFIGURATION_ERROR -> {
-                    status = PLUGIN_CONFIGURATION_ERROR
-                    // As the same error will occur for every file, stop processing
-                    return false
-                }
-
                 null, KtlintRuleEngineWrapper.KtlintResult.Status.NOT_STARTED -> {
                     // File is not a kotlin file
                     return true
@@ -134,6 +134,6 @@ class FormatAction : AnAction() {
             }
         }
 
-        enum class BatchStatus { SUCCESS, PLUGIN_CONFIGURATION_ERROR, FILE_RELATED_ERROR }
+        enum class BatchStatus { SUCCESS, FILE_RELATED_ERROR }
     }
 }