+++ ktlint-plugin/CHANGELOG.md
@@ -4,7 +4,42 @@
 
 ## [Unreleased]
 
-## [0.30.0-beta-2.2025-11-09_18-35-57] - 2025-11-09
+## [0.30.2-beta.2025-12-17_07-09-04] - 2025-12-17
+
+- Add Ktlint ruleset version 1.7.2 by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/777
+
+## [0.30.1] - 2025-12-02
+
+- Fix notification group by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/761
+- Reset Ktlint ".editorconfig" cache on reset of KtlintRuleEngine by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/762
+- Bump org.jetbrains.qodana from 2025.2.1 to 2025.2.2 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/751
+
+## [0.30.0] - 2025-11-23
+
+- Store/read ktlint version in property file for synchronisation with other Ktlint integration by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/744
+- Reconfigure ktlint when switching projects by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/739
+- Assign notification to different notification groups for better customizability by user by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/754
+- Add Ktlint ruleset 1.8.0 by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/755
+- Bump actions/checkout to 5 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/703
+- Bump org.jetbrains.qodana to 2025.2.1 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/706
+- Bump actions/setup-java to 5 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/709
+- Update max IDE build to 253.* by @github-actions[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/714
+- Change platform type from IC to IU by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/716
+- Bump gradle/actions to 5 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/719
+- Bump io.mockk:mockk to 1.14.6 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/720
+- Bump org.jetbrains.kotlinx.kover 0.9.3 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/726
+- Bump Gradle Wrapper to 8.14.3 by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/730
+- Limit the number of IDEA versions that are verified in GitHub build by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/735
+- Bump actions/upload-artifact from 4 to 5 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/736
+- Bump org.junit.jupiter:junit-jupiter 6.0.1 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/741
+- Bump org.junit.platform:junit-platform-launcher 6.0.1 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/742
+- Bump org.jetbrains.intellij.platform to 2.10.4 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/740
+- Bump org.jetbrains.kotlin.jvm to 2.2.21 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/733
+- Bump org.jetbrains.intellij.platform to 2.9.0 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/712
+- Bump org.jetbrains.kotlin.jvm from 2.2.20 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/713
+- Bump org.jetbrains.kotlinx.kover to 0.9.2 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/715
+
+## [[0.30.0-beta-2.2025-11-09]_18-35-57] - 2025-11-09
 
 - Store/read ktlint version in property file for synchronisation with other Ktlint integration by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/744
 - Reconfigure ktlint when switching projects by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/739
@@ -23,14 +58,9 @@
 - Bump org.junit.platform:junit-platform-launcher 6.0.1 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/742
 - Bump org.jetbrains.intellij.platform to 2.10.4 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/740
 - Bump org.jetbrains.kotlin.jvm to 2.2.21 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/733
-- Bump actions/checkout to 5 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/703
-- Bump org.jetbrains.qodana to 2025.2.1 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/706
-- Bump actions/setup-java to 5 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/709
 - Bump org.jetbrains.intellij.platform to 2.9.0 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/712
 - Bump org.jetbrains.kotlin.jvm from 2.2.20 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/713
-- Update max IDE build to 253.* by @github-actions[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/714
 - Bump org.jetbrains.kotlinx.kover to 0.9.2 by @dependabot[bot] in https://github.com/nbadal/ktlint-intellij-plugin/pull/715
-- Change platform type from IC to IU by @paul-dingemans in https://github.com/nbadal/ktlint-intellij-plugin/pull/716
 
 ## [[0.30.0-beta-1.2025-10-11]_14-19-31] - 2025-10-11
 
@@ -488,8 +518,11 @@
 - MVP: plugin that runs ktlint scanner and provides annotations. Lots more to do!
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 
-[Unreleased]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.30.0-beta-2.2025-11-09_18-35-57...HEAD
+[Unreleased]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.30.2-beta.2025-12-17_07-09-04...HEAD
+[0.30.1]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.30.0...v0.30.1
+[0.30.0]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.30.0-beta-2.2025-11-09...v0.30.0
 [0.30.0-dev-1.2025-10-06]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.29.0...v0.30.0-dev-1.2025-10-06
+[0.30.0-beta-2.2025-11-09]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.30.0-beta-1.2025-10-11...v0.30.0-beta-2.2025-11-09
 [0.30.0-beta-1.2025-10-11]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.30.0-dev-1.2025-10-06...v0.30.0-beta-1.2025-10-11
 [0.29.0]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.29.0-beta-1.2025-08-04...v0.29.0
 [0.29.0-beta-1.2025-08-04]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.28.0...v0.29.0-beta-1.2025-08-04
@@ -531,6 +564,7 @@
 [0.2.0]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.0.2...v0.1.0
 [0.0.2]: https://github.com/nbadal/ktlint-intellij-plugin/commits/v0.0.2
+[0.30.2-beta.2025-12-17_07-09-04]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.30.1...v0.30.2-beta.2025-12-17_07-09-04
 [0.30.0-dev-1.2025-10-06_18-30-17]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.29.0...v0.30.0-dev-1.2025-10-06_18-30-17
 [0.30.0-beta-2.2025-11-09_18-35-57]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.30.0-beta-1.2025-10-11...v0.30.0-beta-2.2025-11-09_18-35-57
 [0.30.0-beta-1.2025-10-11_14-19-31]: https://github.com/nbadal/ktlint-intellij-plugin/compare/v0.30.0-dev-1.2025-10-06...v0.30.0-beta-1.2025-10-11_14-19-31